Fix finding of CXX prefix
When configuring the package, the finding of the correct CXX prefix is
crucial (see the file 'tools/config/configure.R'). As the function call
'read_r_config' (and, thus, 'R CMD config') does not return 'NULL' (but
an error message) if a constant is not configured, the wrong prefix is
chosen. This results in compilation errors.

To fix this problem, it needs to be checked whether the call to
'read_r_config' returns an error messsage or not.

Additionally, remove trailing whitespace in 'tools/config/configure.R'.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>