Set working directory for config.R correctly even if setwd is set outside
`setwd` can be set in a `~/.Rprofile` or `Rprofile.site`, to provide
a default path when R opens. This however confuses the config script,
as the R working path might then be different from the system path
from where the script is called.

This patch sets the current R code's working directory to the same
as it was previously assumed, the calling scripts (`configure` or
`cleanup`) folders, through the `PWD` shell env var. After this
the relative paths for the tools folder becomes correct.

Signed-off-by: Gergely Imreh <gergely.imreh@faculty.ai>